compute/requirements.txt
numpy==2.1.3
pika==1.3.2
pika-stubs==0.1.3
PyYAML==6.0.2